更新游戏(尚未完成)、评分模块(尚未完成),新增查询是否点赞、收藏接口。

+++ src/main/java/com/example/gamevibe/model/vo/UserVO.java
@@ -38,6 +38,6 @@ public class UserVO implements Serializable {
     private int fans_count;
 
     @ApiModelProperty(value = "点赞与收藏数", example = "1234")
-    private int like_collect_count;
+    private long like_collect_count;
 
 }
